+++ packages/block-library/src/more/index.js
@@ -15,7 +15,7 @@ import save from './save';
 import transforms from './transforms';
 import { unlock } from '../lock-unlock';
 
-const { fieldsKey } = unlock( blocksPrivateApis );
+const { fieldsKey, formKey } = unlock( blocksPrivateApis );
 
 const { name } = metadata;
 
@@ -46,9 +46,11 @@ if ( window.__experimentalContentOnlyPatternInsertion ) {
 			id: 'customText',
 			label: __( 'Content' ),
 			type: 'richtext',
-			shownByDefault: true,
 		},
 	];
+	settings[ formKey ] = {
+		fields: [ 'customText' ],
+	};
 }
 
 export const init = () => initBlock( { name, metadata, settings } );