+++ packages/block-library/src/verse/index.js
@@ -16,7 +16,7 @@ import save from './save';
 import transforms from './transforms';
 import { unlock } from '../lock-unlock';
 
-const { fieldsKey } = unlock( blocksPrivateApis );
+const { fieldsKey, formKey } = unlock( blocksPrivateApis );
 
 const { name } = metadata;
 
@@ -51,9 +51,11 @@ if ( window.__experimentalContentOnlyPatternInsertion ) {
 			id: 'content',
 			label: __( 'Content' ),
 			type: 'richtext',
-			shownByDefault: true,
 		},
 	];
+	settings[ formKey ] = {
+		fields: [ 'content' ],
+	};
 }
 
 export const init = () => initBlock( { name, metadata, settings } );