+++ packages/block-library/src/video/index.js
@@ -16,7 +16,7 @@ import save from './save';
 import transforms from './transforms';
 import { unlock } from '../lock-unlock';
 
-const { fieldsKey } = unlock( blocksPrivateApis );
+const { fieldsKey, formKey } = unlock( blocksPrivateApis );
 
 const { name } = metadata;
 
@@ -43,7 +43,6 @@ if ( window.__experimentalContentOnlyPatternInsertion ) {
 			id: 'video',
 			label: __( 'Video' ),
 			type: 'media',
-			shownByDefault: true,
 			mapping: {
 				id: 'id',
 				url: 'src',
@@ -59,9 +58,11 @@ if ( window.__experimentalContentOnlyPatternInsertion ) {
 			id: 'caption',
 			label: __( 'Caption' ),
 			type: 'richtext',
-			shownByDefault: false,
 		},
 	];
+	settings[ formKey ] = {
+		fields: [ 'video' ],
+	};
 }
 
 export const init = () => initBlock( { name, metadata, settings } );