+++ packages/global-styles-engine/src/settings/get-setting.ts
@@ -36,6 +36,7 @@ const VALID_SETTINGS = [
 	'custom',
 	'dimensions.aspectRatio',
 	'dimensions.minHeight',
+	'dimensions.width',
 	'layout.contentSize',
 	'layout.definitions',
 	'layout.wideSize',